Update main.py
(Part d): Added function to generate trees of varying depths from 1 to 25 nodes. Added graphing component. 
The assignment instructions said to "add a function in main." I wasn't sure what "in main" meant, so I added it separately and called it from main. Can one of you please move that inside main, because I was unsure of how to do that.